Bug 536324 part 3 - Change consumers of nsIChannel. r=sworkman

+++ embedding/components/webbrowserpersist/src/nsWebBrowserPersist.cpp
@@ -876,7 +876,7 @@ nsWebBrowserPersist::OnDataAvailable(
             }
         }
 
-        int32_t channelContentLength = -1;
+        int64_t channelContentLength = -1;
         if (!cancel &&
             NS_SUCCEEDED(channel->GetContentLength(&channelContentLength)))
         {
